+++ module-api/src/main/java/com/wsws/moduleapi/feed/dto/answer/put_patch/AnswerPatchApiRequest.java
@@ -1,6 +1,6 @@
 package com.wsws.moduleapi.feed.dto.answer.put_patch;
 
-import com.wsws.moduleapplication.feed.dto.answer.AnswerEditServiceRequest;
+import com.wsws.moduleapplication.feed.dto.answer.edit.AnswerEditServiceRequest;
 import jakarta.validation.constraints.NotNull;
 import org.springframework.web.multipart.MultipartFile;
