+++ module-application/src/main/java/com/wsws/moduleapplication/feed/dto/answer/edit/AnswerCreateServiceResponse.java
@@ -1,4 +1,4 @@
-package com.wsws.moduleapplication.feed.dto.answer;
+package com.wsws.moduleapplication.feed.dto.answer.edit;
 
 public record AnswerCreateServiceResponse(
         Long answerId
