+++ module-application/src/main/java/com/wsws/moduleapplication/feed/dto/answer/edit/AnswerEditServiceRequest.java
@@ -1,4 +1,4 @@
-package com.wsws.moduleapplication.feed.dto.answer;
+package com.wsws.moduleapplication.feed.dto.answer.edit;
 
 import org.springframework.web.multipart.MultipartFile;
