+++ module-application/src/main/java/com/wsws/moduleapplication/feed/dto/answer/read/AnswerCommentFindServiceResponse.java
@@ -1,4 +1,4 @@
-package com.wsws.moduleapplication.feed.dto.answer;
+package com.wsws.moduleapplication.feed.dto.answer.read;
 
 import lombok.Builder;
