+++ module-application/src/main/java/com/wsws/moduleapplication/feed/dto/answer/read/AnswerCountByUserServiceResponse.java
@@ -1,4 +1,4 @@
-package com.wsws.moduleapplication.feed.dto.answer;
+package com.wsws.moduleapplication.feed.dto.answer.read;
 
 public record AnswerCountByUserServiceResponse(
         Long answerCount
