+++ module-application/src/main/java/com/wsws/moduleapplication/feed/dto/answer/read/AnswerFindByUserServiceRequest.java
@@ -1,0 +1,11 @@
+package com.wsws.moduleapplication.feed.dto.answer.read;
+
+import java.time.LocalDateTime;
+
+public record AnswerFindByUserServiceRequest(
+        String userId,
+        Long answerId,
+        LocalDateTime cursor,
+        int size
+) {
+}
